Fix WTForms 2.x
It was failing somewhere downstack of DuplicateField, with the error

    TypeError: Must provide one of _form or _meta

Starting in WTForms 2.0, each form has a "form meta" object. The base
Field class wants to grab that and store it[1], so DuplicateField can't do
its dumb trick of passing _form=None to the underlying field any more.
I'm not sure why it was doing that, but afaict none of the builtin
fields even look at the _form parameter, so i doubt it matters.

[1]: https://github.com/wtforms/wtforms/commit/927778c1ac0236c9c914b4b8fff5fa71a7b2b5f1

Updates veekun/spline-pokedex#104.